IMPALA-13350: Fix Workload Management flush_on_interval Test
The test 'flush_on_interval' in test_query_log.py is failing. The
cause is not allowing enough time for the workload management query
processing loop to execute before checking the number of queries
written to the query log table.

The fix is to allow more time for the processing loop to execute.

Change-Id: I2fb1034ca63e170d5e57a6ece9b47da5dafebff4
Reviewed-on: http://gerrit.cloudera.org:8080/21750
Reviewed-by: Riza Suminto <riza.suminto@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>