IMPALA-13322: Fix alter on SystemTables
Fixes SystemTable.load to handle incremental loads - such as during
alter - and correctly update the metastore Table so subsequent
operations see the correct version of the table schema.

Updates QueryScanner to allow unknown columns and log a warning. This is
primarily needed when another coordinator has been upgraded and modified
table metadata with new columns that this coordinator doesn't recognize.

Adds TestQueryLive::test_alter to test that altering the table works and
the table is usable afterwards, including attempting to query unknown
columns as we might encounter them during upgrades.

Change-Id: I9a59e58c086e659941e0db8a2b893ac6dcc5143a
Reviewed-on: http://gerrit.cloudera.org:8080/21792
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>