IMPALA-12686: Switch to toolchain with basic debug info
This switches to a toolchain that has been built with basic debug
information (-g1). This is useful for getting better stack traces
when in library code. The toolchain has also been built with -gz
to compress the debug information.

Some components already built with more debug information (e.g. -g)
and the new toolchain preserves this. This skips adding debug
information for tools like CMake, Mold, etc. It also skips adding
debug information for LLVM's release build. Even at -g1, LLVM's
release build has an enormous amount of debug information, and it
would add hundreds of MBs to impalad's binary size to include it.

This adds about 31MB to the compressed binary size for Impala. It
actually reduces the size of the toolchain by a few hundred MB due
to the compression. However, all libraries now have more debug
information than they did before.

Link commands use a bit more memory than before. The final build
in build-all-flag-combinations.sh tests setting a custom version
for the Java build. Everything is in ccache at that point, so if
it builds the backend tests, there will be many link invocations
running simultaneously, which can overload the system memory.
This modifies that location to use -notests, as it is not testing
the build of backend tests.

Testing:
 - Ran core tests
 - Checked for changes in build time

Change-Id: I7b962c350cc5f1f2b24ca7a52b940ec9e87a7745
Reviewed-on: http://gerrit.cloudera.org:8080/21471
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>