IMPALA-11934: Adjust the timeout interval for temporary file tests
This patch increases the timeout interval for temporary file test
cases. Initially, the timeout was set to 2 seconds, as it was for
hdfs files. However, after integrating ozone for temporary files,
this timeout seems too short for ozone uploads, causing some
related test cases in the ozone builds to become flaky. So, the
timeout is extended to 20 seconds in this patch.

Tests:
Ran disk-io-mgr-test and tmp-file-mgr-test, and no failures were
observed.

Change-Id: I23ec44abd73426ed223aedb5d5634512861ac571
Reviewed-on: http://gerrit.cloudera.org:8080/21835
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>