IMPALA-13406: Switch to curl 8.10.1 to resolve CVEs
This bumps the toolchain build and switches the version
of Curl to 8.10.1. This resolves several medium and low
severity CVEs that are present in Curl 7.78. See
https://curl.se/docs/vuln-7.78.0.html

This also changed the Curl build to stop including
unnecessary features like TELNET/GOPHER/FTP/etc. That
also reduces the surface area for CVEs.

An error message changed slightly in the new Curl
version, so this updates the test to match.

Testing:
 - Ran an exhaustive job

Change-Id: I844578187f1f3f791e4e02d7d686c3e444963806
Reviewed-on: http://gerrit.cloudera.org:8080/21858
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>