IMPALA-13426: Log Java debug sleeps at debug
Updates DebugUtils to log sleeps triggered by debug actions at debug
level, so they're visible in the dev and test environments. Debug
actions are opt-in, and it's important to identify if and when sleeps
happen to ensure tests are working correctly. This updates them to match
the backend, which already logs debug action sleeps at VLOG(1).

Change-Id: Ic5ffec41d62a8b7fb3730e5b3fb7b7dab9ef2972
Reviewed-on: http://gerrit.cloudera.org:8080/21904
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>