IMPALA-13456: Calcite planner: fix issues with quotes
Changed the parser to use quotes that are inline with how Impala
treats quotes, including allowing single quotes, double quotes,
and back ticks for aliases, and also allowing the backslash
to be used as an escape character.  This is inline with what
BigQuery uses in Calcite.

A couple of unit tests were added, but these will be tested more
extensively by the ParserTest frontend unit test when that gets
committed.

Also, added the VALUE as a nonreserved keyword which is used in
the tpcds queries (along with the doublequotes)

Change-Id: I67ebb19912714c240b99a42d9f2f02f78c189350
Reviewed-on: http://gerrit.cloudera.org:8080/21942
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>