IMPALA-13459: Handle duplicate table in same query
WHen there are 2 references of the same table in a query, there needs
to be a unique alias name used within the TableRef object. Code has
been added to generate an alias.

IMPALA-13460 has been filed because we should use the user provided
alias name rather than a generated alias name. This is a little more
difficult to implement because Calcite has a limitation in that their
table object at validation time is equivalent to a FeTable in that there
is only one object for the multiple tables.

In order to fix IMPALA-13460, there is a Calcite bug that has to be
fixed. We'd have to generate our own TableScan object underneath their
LogicalTableScan that would hold an alias. This TableScan can be
generated through their RelBuilder Factory object. But the current
code creates the LogicalTableScan directly rather than go through
a factory, so that would need to be fixed first.

There are no unit tests attached to this Jira, but there are some
tpcds queries that will start working when this gets committed.

Change-Id: Ib9997bc642c320c2e26294d7d02a05bccbba6a0d
Reviewed-on: http://gerrit.cloudera.org:8080/21945
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>