IMPALA-13467: Fix partition list size calculation for empty Iceberg scan nodes
This patch adds a condition that checks whether the IcebergScanNode
contains any files before using the size of the partition list. The
partition list size of Iceberg tables is always one regardless of the
scanned files. This behavior can cause NPE in runtime filter generation.
By setting the calculated partition size to 0, the runtime filter
generation does not occur.

Change-Id: I5a0595831f3bd87074144ab7d5da27508e73ef33
Reviewed-on: http://gerrit.cloudera.org:8080/21964
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>