IMPALA-13976: Switch from boost::function/bind/mem_fn to std
Impala's code uses boost::function/bind/mem_fn in several places, but
all of this has been incorporated into the standard library. When
trying to upgrade to newer Boost (1.91.0), the use of boost::bind
and other functionality created conflicts with the standard library.
Given that there is no real advantage to staying with the Boost
version of these, this switches all locations to the standard
library.

One difference is that boost::function has a different API from
std::function. boost::function uses empty() to determine if it is
null whereas std::function can simply be compared to null. Overall,
these differences are minor and this is simply a boilerplate change.

This leaves some uses of boost::function in udf-test-harness.h,
as that file claims to support C++98. These can be removed when
we move that to a more recent C++ version.

Testing:
 - Ran a core job

Change-Id: I769d84b02fbe691b3bf2a6d3f52836d5b684bae0
Reviewed-on: http://gerrit.cloudera.org:8080/21967
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Yida Wu <wydbaggio000@gmail.com>
Tested-by: Michael Smith <michael.smith@cloudera.com>