IMPALA-13471: test_enable_reading_puffin() seems to fail in the Ozone build
The tests
TestIcebergWithPuffinStatsStartupFlag::test_[dis|en]able_reading_puffin
queried an Iceberg table that is created during normal dataload from
existing non-filesystem-specific metadata and data files. Therefore the
path of the Puffin stats file that is present in the metadata.json file
does not contain any filesystem-specific prefix, for which Puffin
reading does not work on Ozone.

Note that reading Puffin stats for tables that are created normally do
work on Ozone.

This change modifies the tests to create the table on the fly, modifying
the file path to include the filesystem-specific prefix.

Change-Id: I7afec1c70d7b43bae98289d65749b01ca720e7f7
Reviewed-on: http://gerrit.cloudera.org:8080/22008
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>