IMPALA-13494: Calcite planner group_concat with distinct failing
The following query is failing.

select sum(len_orderkey), sum(len_comment)
from (
  select
    length(group_concat(distinct cast(l_orderkey as string))) len_orderkey,
    length(group_concat(distinct(l_comment))) len_comment
    from tpch.lineitem
    group by l_comment
  ) v

There is code where in AggregationFunction for group_concat that calls
to ignore an implicit cast. The 'isImplicitCast_' member is being used
directly in this function, but the variable is overridden in the
isImplicitCast method for the Calcite planner. A small change was needed
to call the isImplicitCast() function rather than use the member variable.

Change-Id: Idec41597b40a533bc0774b4ff2ab5059c7f324e2
Reviewed-on: http://gerrit.cloudera.org:8080/22025
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>